requirements.txt
dill==0.3.6
mlflow==2.1.1
numpy==1.23.5
pandas==1.5.3
pydantic==1.10.5
PyYAML==6.0
scikit-learn==1.2.1
streamlit==1.18.1
xgboost==1.7.4